+++ graphql_api/tests/rupture_set/conftest.py
@@ -1,3 +1,5 @@
+import datetime
+
 import pytest
 
 
@@ -64,3 +66,65 @@ def create_rs_mutation():
         }
     }
 '''
+
+
+@pytest.fixture()
+def rupture_generation_task(graphql_client, create_gt_mutation, create_rs_mutation):
+
+    # create the GT to be referenced in the AT
+    gt1 = graphql_client.execute(
+        create_gt_mutation,
+        variable_values=dict(
+            created=datetime.datetime.now(datetime.UTC), argument_lists=[dict(k="swept_arg", v=["A", "B"])]
+        ),
+    )
+    print(gt1)
+    gt_id = gt1['data']['create_general_task']['general_task']['id']
+
+    # Now create the Rupture Generation task
+    executed = graphql_client.execute(
+        create_rs_mutation,
+        variable_values=dict(
+            created=datetime.datetime.now(datetime.UTC), gt_id=gt_id, arguments=dict(k="swept_arg", v="A")
+        ),
+    )
+    task_result = executed['data']['create_rupture_generation_task']['task_result']
+    yield task_result
+
+
+@pytest.fixture()
+def create_rupture_set_mutation():
+    yield """
+        mutation (
+            $md5_digest: String!, 
+            $file_name: String!, 
+            $file_size: BigInt!, 
+            $produced_by: ID!
+            $arguments: [KeyValuePairInput],
+            $metrics: [KeyValuePairInput],
+            $created: DateTime!
+        ) {
+              create_rupture_set(
+                input: {
+                  md5_digest: $md5_digest
+                  file_name: $file_name
+                  file_size: $file_size
+                  produced_by: $produced_by
+                  arguments: $arguments
+                  metrics: $metrics
+                  created: $created
+                  }
+              ) {
+              rupture_set { 
+                id
+                file_name
+                file_size
+                md5_digest
+                created
+                produced_by { id __typename, general_task_id}
+                arguments { k v }
+                metrics { k v }
+                }
+              }
+            }
+    """
